refactor: extract _FunctionLowerer into codegen_lowerer
Split the per-kernel scalar/expression lowering out of codegen.py. Move the
shared low-level primitives (the primitive-type map, CodegenError, _type_name)
and the ~680-line _FunctionLowerer visitor -- one contiguous block -- into a
new codegen_lowerer module, leaving codegen.py focused on pipeline/tick
emission (emit_llvm_ir).

The new module imports only from .ast, .codegen_intrinsics, and .utils, so it
does not depend on codegen.py -- no import cycle. codegen.py imports the four
names back and drops four imports the moved code took with it.

Emitted IR is byte-for-byte identical (tests/test_golden_ir.py passes
unchanged); ruff and mypy clean; full suite green.

codegen.py: 2845 -> 2150 lines; new codegen_lowerer.py: 736 lines.
